+++ front-end/machinapp/src/pages/Sitios/Ambientes.jsx
@@ -1,5 +1,5 @@
 import React from 'react'
-import BuscarMaquinas from '../components/organisms/ListarMaquinas'
+import BuscarMaquinas from '../../components/organisms/ListarMaquinas'
 
 const Ambientes=()=>  {
   return (
