+++ front-end/machinapp/src/pages/Sitios/Sedes.jsx
@@ -1,8 +1,8 @@
 import { useEffect, useState } from 'react'
-import BuscarAreas from '../components/organisms/ListarAreas'
-import Layout from '../components/template/Layout.jsx'
+import BuscarAreas from '../../components/organisms/Sitios/ListarAreas.jsx'
+import Layout from '../../components/template/Layout.jsx'
 import { useParams } from "react-router-dom"
-import { axiosCliente } from "../service/api/axios.js"
+import { axiosCliente } from "../../service/api/axios.js"
 
 const Sedes = () => {
   const { idSede } = useParams()
