+++ front-end/machinapp/src/pages/Sitios/Sitios.jsx
@@ -1,5 +1,5 @@
-import BuscarSedes from '../components/organisms/ListarSedes'
-import Layout from '../components/template/Layout'
+import BuscarSedes from '../../components/organisms/Sitios/ListarSedes'
+import Layout from '../../components/template/Layout'
 
 const Sitios=()=>  {
   return (
