+++ front-end/machinapp/src/router/Router.jsx
@@ -12,20 +12,20 @@ import { PerfilRoutes } from "./rutas/PerfilRoute";
 import Fichas from "../pages/Fichas";
 import Historial from "../pages/Historial";
 import Maquinas from "../pages/Maquinas";
-import Sitios from "../pages/Sitios";
+import Sitios from "../pages/Sitios/Sitios";
 /* import PanelControl from "../pages/PaneldeControl"; */
-import Sedes from "../pages/Sedes";
-import Areas from "../pages/Areas";
+import Sedes from "../pages/Sitios/Sedes";
+import Areas from "../pages/Sitios/Areas";
 import { AdminRoute } from "./rutas/AdminRoutes";
 import MaquinasAmbiente from "../pages/MaquinasAmbiente";
 import InfoMaquina from "../pages/InfoMaquina";
 import { ViewFormFichaTecnica } from "../pages/fichastecnicas/ViewFormFichaEquipos";
 import { ViewFormFicha_De_mantenimiento } from "../pages/fichastecnicas/ViewFormFicha_de_mantenimiento";
 import { FichaSolicitudPage } from "../pages/fichastecnicas/ViewFormFichaSolicitudPage";
-import InfoAmbiente from "../pages/InfoAmbiente";
-import InfoSede from "../pages/InfoSede";
-import AreasGeneral from "../pages/AreasGeneral";
-import AmbientesGeneral from "../pages/AmbientesGeneral";
+import InfoAmbiente from "../pages/Sitios/InfoAmbiente";
+import InfoSede from "../pages/Sitios/InfoSede";
+import AreasGeneral from "../pages/Sitios/AreasGeneral";
+import AmbientesGeneral from "../pages/Sitios/AmbientesGeneral";
 import { ResetPassword } from "../pages/auth/ResetPassword";
 
 export const AppRouter = () => {
